Fix network configuration
The old network configuration would not wait for the eth0 interface to
be fully configured. An interface without an IP address can't serve as
an endpoint for a VXLAN tunnel, so as a result the vxlan-* interfaces
would never come up, Upstart would never emit the static-network-up
event, and cloud-init-nonet would stall for 120s waiting for a network
to come up that never would.

Move the creation of the VXLAN interfaces to the post-up scripts for
eth0. Then, have the VXLAN interfaces wait until eth0 is fully
configured.

Fixes hastexo/orchestration#2.